requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
pytz==2024.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.2.1
